Fix for GCC sign-compare error for AzCore, AzFramework ( [-Werror=sign-compare] )

+++ Code/Framework/AzFramework/AzFramework/Asset/AssetSystemComponentHelper.cpp
@@ -114,7 +114,7 @@ namespace AzFramework
 #if AZ_TRAIT_DENY_ASSETPROCESSOR_LOOPBACK
                     bool isIllegalLoopBack = ip == "127.0.0.1";
 #endif
-                    for (int i = 0; isNumeric && i < iplen; ++i)
+                    for (size_t i = 0; isNumeric && i < iplen; ++i)
                     {
                         if (ip[i] == '.')
                         {
